requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
replicate==0.8.1
requests==2.31.0
python-multipart==0.0.6